# api/openapi.yaml
openapi: "3.0.3"
info:
  title: DeFi Yield Hub API
  description: >
    封装 LI.FI Earn API 的 DeFi 收益工具集。
    支持 Vault 发现、持仓查询、跨链存款交易构建。
    可作为通用 Agent Skill 供 Claude、OpenClaw、LangChain 等框架直接调用。
  version: "1.0.0"
  contact:
    name: DeFi Mullet Hackathon
    url: https://lifi.notion.site/defi-mullet-hackathon-1-builder-edition
servers:
  - url: http://localhost:3000
    description: 本地开发环境
  - url: https://your-api-domain.com
    description: 生产环境（部署后替换）
...
paths:
  /vaults:
    get:
      operationId: getVaults
      summary: 查询收益 Vault 列表
      description: >
        返回可用的 DeFi 收益 Vault，支持按链、代币符号、协议、最低 APY、最低 TVL 过滤，
        以及按 APY 或 TVL 排序。支持分页，共有 672+ 个 Vault。
      tags: [Vaults]
      parameters:
        - name: chainId
          in: query
          description: 按链 ID 过滤（例如 42161=Arbitrum, 8453=Base, 1=Ethereum）
          schema:
            type: integer
          example: 42161
        - name: asset
          in: query
          description: 按底层代币符号过滤（例如 USDC、USDT、ETH）
          schema:
            type: string
          example: USDC
        - name: protocol
          in: query
          description: 按协议名过滤（例如 morpho-v1、aave-v3、euler-v2）
          schema:
            type: string
          example: morpho-v1
        - name: minApy
          in: query
          description: 最低总 APY 门槛（百分比，例如 5 表示 5%）
          schema:
            type: number
          example: 5
        - name: minTvl
          in: query
          description: 最低 TVL（USD）
          schema:
            type: number
          example: 1000000
        - name: sortBy
          in: query
          description: 排序字段
          schema:
            type: string
            enum: [apy, tvl, apy7d, apy30d]
          example: apy
        - name: limit
          in: query
          description: 每页数量，默认 20，最大 100
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: cursor
          in: query
          description: 分页游标（来自上一页响应的 nextCursor）
          schema:
            type: string
      responses:
        "200":
          description: 成功返回 Vault 列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultsResponse"
        "500":
          $ref: "#/components/responses/Error"
  /vaults/{chainId}/{address}:
    get:
      operationId: getVaultDetail
      summary: 获取单个 Vault 详情
      description: 根据链 ID 和 Vault 合约地址获取完整的 Vault 信息，包括 APY 历史、TVL、协议信息等。
      tags: [Vaults]
      parameters:
        - name: chainId
          in: path
          required: true
          description: 链 ID
          schema:
            type: integer
          example: 42161
        - name: address
          in: path
          required: true
          description: Vault 合约地址（0x 格式）
          schema:
            type: string
          example: "0x1234567890abcdef1234567890abcdef12345678"
      responses:
        "200":
          description: 成功返回 Vault 详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vault"
        "500":
          $ref: "#/components/responses/Error"
...
components:
  schemas:
    UnderlyingToken:
      type: object
      properties:
        symbol:
          type: string
          example: USDC
        address:
          type: string
          example: "0xaf88d065e77c8cC2239327C5EDb3A432268e5831"
        decimals:
          type: integer
          example: 6
    VaultAnalytics:
      type: object
      properties:
        apy:
          type: object
          properties:
            base:
              type: number
              nullable: true
              example: 5.2
            reward:
              type: number
              nullable: true
              example: 1.1
            total:
              type: number
              nullable: true
              example: 6.3
        apy1d:
          type: number
          nullable: true
        apy7d:
          type: number
          nullable: true
        apy30d:
          type: number
          nullable: true
        tvl:
          type: object
          properties:
            usd:
              type: string
              description: TVL 字符串（非数字，需手动 parseFloat）
              example: "45000000"
    Vault:
      type: object
      properties:
        address:
          type: string
          description: Vault 合约地址，存款时作为 Composer 的 toToken 参数
        chainId:
          type: integer
        name:
          type: string
        protocol:
          type: object
          properties:
            name:
              type: string
        underlyingTokens:
          type: array
          items:
            $ref: "#/components/schemas/UnderlyingToken"
        tags:
          type: array
          items:
            type: string
        analytics:
          $ref: "#/components/schemas/VaultAnalytics"
        isTransactional:
          type: boolean
          description: 是否支持通过 Composer 存款
        isRedeemable:
          type: boolean
    VaultsResponse:
      type: object
      properties:
        vaults:
          type: array
          items:
            $ref: "#/components/schemas/Vault"
        nextCursor:
          type: string
          description: 下一页游标，为空时表示已是最后一页
          nullable: true
        total:
          type: integer
...
  responses:
    Error:
      description: 服务端错误
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
...